\name{sequences-package}
\alias{sequences-package}
\alias{sequences}
\docType{package}

\title{
  A short demo package featuring biological sequences.
}

\description{
  Dummy package used in an R course to illustrate OO programming
  and package development. The course is aimed at bioinformaticians
  and biologists. Relevant illustrative examples used in the
  package are generic sequences as a top virtual class and specific
  biological (DNA and RNA) sequences. 
}

\details{
  \tabular{ll}{
    Package: \tab sequences\cr
    Type: \tab Package\cr
    License: \tab GPL-3\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Laurent Gatto <lg390@cam.ac.uk>
  Robert Stojnic <rs550@cam.ac.uk>
  Maintainer: Laurent Gatto <lg390@cam.ac.uk>
}

\keyword{ package }
